Remove the marker of menu <ul> element
This is a problem in Firefox because Firefox displays markers(bullet
points) of `<li>` element just before the inline contents of the
elements. But Chrome displays them before the element, not depending on
the content of the `<li>` element. So it wasn't visible on Chrome but
it's in Firefox. This CSS change will remove that in Firefox too.