Fix broken footnote anchor link scrolling
Footnotes result in anchor link hrefs such as "#fn:1". However, it appears that
colons are special characters that break the `document.querySelector` function. When
footnotes were clicked, this error was produced: "Failed to execute
'querySelector' on 'Document': '#fn:1' is not a valid selector". The solution implemented
is based on the stackoverflow answer linked below.

https://stackoverflow.com/a/51396346/300212