Add option to show link in meta data for editing posts
The idea was inspired by the [stapelberg theme].

If `editPostUrl` is set, its value will be used to create a link for every post
is displayed in the meta data of the post (if meta not hidden). This is useful
for providing the reader with an easy way to find the proper place to suggest
changes. An example would be an edit prefix for the proper git repo, e.g.:
```
Params:
  editPostUrl: "https://github.com/adityatelange/hugo-PaperMod/edit/exampleSite"
```
For the post saved in `content/posts/markdown-syntax.md`, this will lead to an
href of
"https://github.com/adityatelange/hugo-PaperMod/edit/exampleSite/content/posts/markdown-syntax.md"

The text for the edit button defaults to "Edit" or a translation if available
but can be changed individually by setting the parameter `editPostText`.

This closes #238.

[stapelberg theme]: https://github.com/stapelberg/hugo/blob/d0b2a3edb61a855d87debc3d974ed250853aa931/themes/stapelberg/layouts/_default/single.html#L46