Add Examples section to Readme.md
First of all - thanks you for PaperMod theme!
I built my first static website with your theme and that was a great experience!

About PR:
While I setted up my site with Hugo PaperMod theme I had a lot of problems, because I was a newbie with Hugo.
So I learned a lot not only from wiki pages, but mostly from your example site. 
And I think that it can be helpful to everyone to have more examples of using this theme. 
Thats why I propose to add 'Examples' section to main page of this repository.