Use theme color behind codeblock borders
The PaperMod theme uses a border-radius for codeblocks by default. The
div that contains the corresponding code element uses the default
background color, independent of the selected theme (dark vs. light).
This causes an incorrect coloring in the corners of code blocks.

This commits overwrites the background color of the surrounding div with
the current theme color.