Fix border radius when using line numbers
When using line numbers with table style, the border radius for the
corresponding codeblocks leads to annoying gaps between the line numbers
and the actual code.

This commit removes the border radius when the corresponding element is
displayed inside a table (which is the case when line numbers are used
with table format).