Add support for displaying social icons in footer
This commit adds the following new options to the Sites params:

* SocialIconsFooter
* SocialIconsFooterHome
* HideSocialIconsFooterSections:

The first one can be used to display social icons within the footer for
all pages execpt the homepage. The second one can be set to true to also
display social icons within the homepage. The last one is a list and can
be used to hide social icons on specific sections.

Furthermore, each page can include the page param HideSocialIconsFooter
to prevent social icons from being displayed within the footer.