Add option to add word count to metadata. Closes #735
Reading time is subjective and relative. Word count is absolute and a
more reliable and traditional proxy to estimate how long a post is. It's
not perfect, and there are variations across languages, hence an option
would leave it to users to decide whether it's relevant to their use
cases.

Introduce a ShowWordCount option and use it in post_meta.html.

Hugo supports word counts natively, c.f.
https://gohugo.io/variables/page/#page-variables