Modify breadcrumbs logic
First part of the breadcrumbs generation algorithm is to remove the
base URL from the parent's permalink URL.

In the existing code, this is achieve by replacing all instances of
base URL in the parent's permalink URL by an empty string.

This algorithm fails to produces breadcrumbs in the instance where the
base URL is "/" for example.

I believe this is erroneous as only the first instance of the base URLi,
starting at the beginning of the parent's permalink URL should be
replaced by an empty string.

This is the reason I replaced the call to "replace" by a call to
"replaceRE".