Use version appropriate Go module path
This theme is packaged as a Go module in order to allow it to be managed as a dependency of Hugo-based websites.

Due to the current use of a tag name format that does not meet the the semver-compliant requirement of the Go module
framework, the standard procedure for installing the theme in a Hugo-based website as a Go module dependency results in
the beta version of the theme from the tip of the theme repository's default branch being installed.

The resulting "pseudo-version" of the dependency makes it so that updates of the dependency cannot be managed
automatically via the popular Dependabot service.

A Go module is identified by a "module path". The Go module framework requires that this path have a major version
suffix if the project is at a version of 2.0.0 or above.

This project is at version 7.0, so its Go module path must have a suffix of v7 in order to be compliant with the Go
module framework.

Using a compliant module path as well as compliant tag names benefits the users of this theme by allowing them to manage
the dependency in a safe and efficient manner.